python-requirements.txt
django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
coreapi==2.3.3
coreapi-cli==1.0.9
pyaml==19.12.0
drf-extensions==0.6.0
gunicorn==20.0.4
psycopg2==2.8.4
Markdown
django_extensions
django-shell-plus
